requirements.txt
pandas==1.5.3
numpy==1.24.2
pyspark==3.4.0
delta-spark==2.4.0
kafka-python==2.0.2